GFX Entity handler docs+minor refactoring
